core: suggest a new skeleton script template
Handling argument and command parsing in Bash is really difficult, and I
am interested in proposing a new skeleton script that is primarily aimed
at handling argument parsing for various types of commands we will likely
want to implement in the future. My goal is that this skeleton has an
easily-extensible structure to accept new commands easily and that the
argument parsing is robust and easy to reuse while being as easy to read
as possible.

It's my goal that we can port the existing functionality to
this script with little effort and have a good foundation for expanding
features without having to go through detailed back-and-forth about how
to implement the boilerplate for argument parsing in a piece-by-piece
manner. This will likely only result in us having a cumbersome script in
need of large refactoring down the line.

I designed the argument parser to accept long and short arguments both
with and without values.

I designed the skeleton to support and show quality examples of commands
including commands that have arguments and subcommands.

The following can be used to set the namespace to 'my-ns', for example.
   skeleton.sh -nmy-ns
   skeleton.sh -n my-ns
   skeleton.sh --namespace=my-ns
   skeleton.sh --namespace my-n

Or, this would enable verbose mode:
   skeleton.sh -v
   skeleton.sh --verbose

I implemented -h/--help flags and a 'help' command.

I implemented a placeholder 'ceph' command to run a ceph CLI command
with arbitrary args.

I implemented a placeholder 'status' command that supports a -l/--long
argument.

I implemented a placeholder 'operator' command with an additional
'restart' subcommand.

The skeleton also throws errors on common corner cases like
- a flag's value is another flag
- a flag's value is empty
- a flag isn't supported
- a command isn't supported
- a command wasn't given
- extraneous commands were given at the end of input

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>