remove controller-runtime client
Because this is not an operator, we don't need to use the help
controller-runtime provides. Additionally, all code was using the
Kubernetes clientset, and a lesson-learned in Rook is that unit testing
multiple types of clients gets to be a burden for complex code. Let's
keep the number of client interfaces to a minimum.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>