Use recommended context.Context passing
Documentation in context.Context states never to embed a context.Context
in a struct. It should always be passed to functions and methods,
ideally as the first parameter. Kubernetes' Go API follows this syntax,
and since it is much easier to fix the issue in kubectl-rook-ceph
compared to in rook/rook, we should do so.

The old k8sutil.Context has been renamed to k8sutil.Clientsets to
reflect its new purpose.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>