core: fix osd hang during purge-osd
with earlier approach purging an OSD was hanging.
since, with bash implementation it was passing so now
we have two to run command inside pod `execCmdInPodWithoutReturn`
which will directly print the command in shell and another way
is `execCmdInPodWitReturn` which will first wait for command to
finish and then return the output.

Signed-off-by: subhamkrai <srai@redhat.com>