namespace: namespace flags are not working
cluster namespace and operator namespace needs
better handling. It was throwing an error and was exited
when namespace args are passed.
Now, the default cluster ns is rook-ceph, and for
operator it is empty but we check later in the code
that if CephClusterNamespace != "" && OperatorNamespace == ""
then assign the cluster ns to operator ns.

Signed-off-by: subhamkrai <srai@redhat.com>