clean: ensure cephcluster cleanup job runs
The cleanup policy was not allowed to take effect during the
cluster cleanup. The finalizer was immediately be removed instead
of remaining until the operator has a chance to handle the cleanup
policy and internally remove the finalizer. Now the cleanup policy
is allowed to be applied and the cleanup job(s) will be started
to cleanup the host path and disks from OSDs.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>