csi: look for subvolumepath instead of subvolumename
In case of ROX pvc, the subvolume name is different and is not listed
in the backend as it is just a shallow volume and it links to the source
path.
incase the source volume is delete the subvolme will be shown stale, but
actually that is being used by the ROX pvc.
Soinstead getting the subvolumename to look for stale subvolumes,
it is better to retrive the name from the subvolumepath.

Signed-off-by: yati1998 <ypadia@redhat.com>